+++ colors.py
@@ -1,0 +1,13 @@
+WHITE = (255, 255, 255)
+GREY = (128, 128, 128)
+BLUE = (0, 0, 255)
+BLUE_NEGG = (0, 128, 255)
+YELLOW = (255, 255, 0)
+GREEN = (0, 255, 0)
+RED = (255, 0, 0)
+PURPLE = (255, 0, 255)
+BLACK = (0, 0, 0)
+ORANGE = (255, 165, 0)
+PINK = (255, 105, 180)
+CYAN = (0, 255, 255)
+BROWN = (200, 165, 155)
